Extract gas estimation into separate module
...and make it more efficient
